move optimize atop code to top.py
